+++ runtime/src/startup/mod.rs
@@ -1,7 +1,7 @@
 //! Runtime components related to process startup.
 
 use crate::TockSyscalls;
-use libtock_platform::Termination;
+use libtock_platform::{syscall_class, RawSyscalls, Termination};
 
 // Include the correct `start` symbol (the program entry point) for the
 // architecture.
@@ -62,17 +62,48 @@ pub fn handle_main_return<T: Termination>(result: T) -> ! {
     Termination::complete::<TockSyscalls>(result)
 }
 
+// The runtime header, which is generated by the linker script and placed at the
+// beginning of the app binary.
+#[repr(C)]
+struct RtHeader {
+    start: usize,
+    initial_break: *mut (),
+    stack_top: *mut (),
+    data_size: usize,
+    data_flash_start: *const u8,
+    data_ram_start: *mut u8,
+    bss_size: usize,
+    bss_start: *mut u8,
+}
+
 // rust_start is the first Rust code to execute in the process. It is called
 // from start, which is written directly in assembly.
 #[no_mangle]
 extern "C" fn rust_start() -> ! {
-    // TODO: Call memop() to inform the kernel of the stack and heap sizes +
-    // locations. Also, perhaps we should support calling a heap initialization
-    // function?
-
     extern "Rust" {
         fn libtock_unsafe_main() -> !;
+        static rt_header: RtHeader;
     }
+
+    // TODO: Implement a safe memop API in libtock_platform and migrate these
+    // calls to that API.
+    //
+    // Safety: Memop operations 10 and 11 are always memory-safe, as they do not
+    // impact the execution of this process.
+    #[cfg(not(feature = "no_debug_memop"))]
+    unsafe {
+        TockSyscalls::syscall2::<{ syscall_class::MEMOP }>([
+            10u32.into(),
+            rt_header.stack_top.into(),
+        ]);
+        TockSyscalls::syscall2::<{ syscall_class::MEMOP }>([
+            11u32.into(),
+            rt_header.initial_break.into(),
+        ]);
+    }
+
+    // Safety: libtock_unsafe_main is defined by the set_main! macro, and its
+    // signature matches the signature in the `extern` block in this function.
     unsafe {
         libtock_unsafe_main();
     }
