+++ runtime/src/startup/start_prototype.rs
@@ -9,17 +9,8 @@
 // hand. This is a Rust version of `start` for developers who are working on
 // `start`.
 
-#[repr(C)]
-struct RtHeader {
-    start: usize,
-    initial_break: *mut (),
-    stack_top: usize,
-    data_size: usize,
-    data_flash_start: *const u32,
-    data_ram_start: *mut u32,
-    bss_size: usize,
-    bss_start: *mut u8,
-}
+use super::RtHeader;
+use core::arch::asm;
 
 #[link_section = ".start"]
 #[no_mangle]
@@ -30,7 +21,7 @@ extern "C" fn start_prototype(
     _app_break: usize,
 ) -> ! {
     use crate::TockSyscalls;
-    use libtock_platform::{RawSyscalls, syscall_class};
+    use libtock_platform::{syscall_class, RawSyscalls};
 
     let pc: usize;
     unsafe {
@@ -43,20 +34,23 @@ extern "C" fn start_prototype(
         // Binary is in an incorrect location: report an error via
         // LowLevelDebug then exit.
         unsafe {
-            TockSyscalls::syscall4::<syscall_class::COMMAND>([
-                8 as *mut (),
-                1 as *mut (),
-                2 as *mut (),
-                0 as *mut (),
+            TockSyscalls::syscall4::<{ syscall_class::COMMAND }>([
+                8u32.into(),
+                1u32.into(),
+                2u32.into(),
+                0u32.into(),
             ]);
-            TockSyscalls::syscall2::<syscall_class::EXIT>([0 as *mut (), 0 as *mut ()]);
+            TockSyscalls::syscall2::<{ syscall_class::EXIT }>([0u32.into(), 0u32.into()]);
         }
     }
 
     // Set the app break.
     // TODO: Replace with Syscalls::memop_brk() when that is implemented.
     unsafe {
-        TockSyscalls::syscall2::<syscall_class::MEMOP>([0 as *mut (), rt_header.initial_break]);
+        TockSyscalls::syscall2::<{ syscall_class::MEMOP }>([
+            0u32.into(),
+            rt_header.initial_break.into(),
+        ]);
     }
 
     // Set the stack pointer.
@@ -70,8 +64,8 @@ extern "C" fn start_prototype(
     // Copy .data into place. Uses a manual loop rather than
     // `core::ptr::copy*()` to avoid relying on `memcopy` or `memmove`.
     let mut remaining = rt_header.data_size;
-    let mut src = rt_header.data_flash_start;
-    let mut dest = rt_header.data_ram_start;
+    let mut src = rt_header.data_flash_start as *const u32;
+    let mut dest = rt_header.data_ram_start as *mut u32;
     while remaining > 0 {
         unsafe {
             core::ptr::write(dest, *(src));
